Delete LifecycleExecutor class and its test
